finish course drive

+++ src/store/drive/driveApi.js
@@ -8,17 +8,45 @@ import {
   doc,
   deleteDoc,
   updateDoc,
+  documentId,
+  getDoc,
 } from "firebase/firestore";
-import { db } from "../../app/firebase";
-import { collectIdsAndDocs } from "../../common/utils";
+import { db, storage } from "../../app/firebase";
+import { collectIdsAndDocs, generateUUID } from "../../common/utils";
 import { fileConverter } from "../converters";
-import { hasDuplicateName } from "./driveUtils";
+import { hasDuplicateName, generateFileName } from "./driveUtils";
+import {
+  ref,
+  uploadBytes,
+  deleteObject,
+  getDownloadURL,
+} from "firebase/storage";
+
+export const getMarkdownFileDownloadUrlBelongToCourse = async (
+  courseId,
+  id
+) => {
+  const file = await getFileById(id);
+  if (file.courseId !== courseId)
+    throw new Error("The file does not belong to this course!");
+  if (file.type !== "text/markdown")
+    throw new Error(" This is not a file could be previewed!");
+  return await getFileDownloadURL(file.url);
+};
+
+export const getFileById = async (id) => {
+  const docRef = doc(db, "files", id).withConverter(fileConverter);
+  const docSnap = await getDoc(docRef);
+  if (!docSnap.exists()) throw new Error("File isn't exist");
+  return collectIdsAndDocs(docSnap);
+};
 
 export const getFiles = async (courseId) => {
   const collectionRef = collection(db, "files").withConverter(fileConverter);
   const q = query(collectionRef, where("course_id", "==", courseId));
   const querySnapshot = await getDocs(q);
-  const files = querySnapshot.docs.map(collectIdsAndDocs);
+  const files = {};
+  querySnapshot.docs.forEach((doc) => (files[doc.id] = collectIdsAndDocs(doc)));
   return files;
 };
 
@@ -31,21 +59,59 @@ export const renameFileOrFolder = async (newName, data, dict) => {
   await updateDoc(docRef, { name: newName });
 };
 
-export const deleteFileOrFolder = async (id) => {
+export const deleteFileOrFolder = async (data) => {
+  const { id, isDirectory, children, url } = data;
+  if (isDirectory && children && children.length > 0) {
+    throw new Error("Folder must be empty before delete");
+  }
+  if (!isDirectory) await deleteFile(url);
   const docRef = doc(db, "files", id);
   await deleteDoc(docRef);
+};
+
+const deleteFile = async (url) => {
+  const desertRef = ref(storage, url);
+  await deleteObject(desertRef);
 };
 
 export const createFolder = async (data) => {
   const collectionRef = collection(db, "files").withConverter(fileConverter);
   const { id } = await addDoc(collectionRef, data);
-
-  const floderData = {
-    id,
-    ...data,
-  };
-
-  return floderData;
+  return id;
 };
 
-export const uploadFile = async () => {};
+export const uploadFile = async (file, courseId, parent, dict) => {
+  if (file.size > 1024 * 1024 * 8)
+    throw new Error("Couldn't upload file greater than 8Mb!");
+
+  const name = generateFileName(file.name, parent, dict);
+  const index = name.lastIndexOf(".");
+  const suffix = index === -1 ? "" : `.${name.substr(index + 1)}`;
+  const uuid = generateUUID(file);
+  const storageRef = ref(storage, `${courseId}/${uuid}${suffix}`);
+  const {
+    metadata: { contentType, size, timeCreated, updated, fullPath },
+  } = await uploadBytes(storageRef, file);
+  const data = {
+    courseId,
+    createdAt: new Date(timeCreated),
+    updatedAt: new Date(updated),
+    isDirectory: false,
+    name,
+    url: fullPath,
+    parentId: parent.id,
+    size,
+    type: contentType,
+  };
+
+  const collectionRef = collection(db, "files").withConverter(fileConverter);
+  const { id } = await addDoc(collectionRef, data);
+  data.id = id;
+  return data;
+};
+
+export const getFileDownloadURL = async (url) => {
+  const storageRef = ref(storage, url);
+  const downloadURL = await getDownloadURL(storageRef);
+  return downloadURL;
+};
